pyproject.toml
[tool.poetry]
name = "nzshm-runzi"
version = "0.1.0"
description = "scripting and cli for the NSHM"
authors = ["GNS NSHM programme"]
license = "GPL-3.0-only"
readme = "README.md"
packages = [
    {include = "runzi"},
    {include = "scripts"}
]

[tool.poetry.scripts]
runzi = 'scripts.cli:rnz'

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
gitpython = "3.1.11"
py4j = "0.10.9.1"
pytz = "^2022.7"
nshm-toshi-client = ">=1.0.0"
prompt-toolkit = "3.0.20"
termcolor = "1.1.0"
pyfiglet = "0.8.post1"
boto3 = "^1.26.43"
lxml = "^4.9.2"
tqdm = "^4.65.0"
toml = "^0.10.2"
click = "^8.1.7"
python-dateutil = "^2.8.2"

# [tool.poetry.extras]
# openquake = ["openquake-engine", "fiona"]
# converter =["openquake-converters-ucerf"]
nzshm-common = {extras = ["geometry"], version = "^0.8.4"}
toshi-hazard-store = {git = "https://github.com/GNS-Science/toshi-hazard-store.git", rev = "update-nzshm-model"}
nzshm-model = {git = "https://github.com/GNS-Science/nzshm-model.git", rev = "pre-release"}
inquirer = "^3.4.0"
pygments = "^2.19.1"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.3.4"
flake8 = ">=7.1.1"
black = ">=25.1.0"
isort = ">=6.0.0"
bump2version = "^1.0.1"
tox = "^4.24.1"
mypy = "^1.14.1"
pytest-cov = "^6.0.0"
types-requests = "^2.32.0.20241016"
types-pytz = "^2025.1.0.20250204"
types-python-dateutil = "^2.9.0.20241206"



[tool.poetry.group.doc.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.5.50"
mkdocs-click = "^0.8.1"
mkdocs-include-markdown-plugin = "^7.1.2"
mkdocstrings = "^0.27.0"


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
# you can skip files as below
#skip_glob = docs/conf.py

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py312', 'py311', 'py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | runzi/CONFIG
)/
'''

[tool.mypy]
ignore_missing_imports = true
exclude = [
    'runzi/CONFIG',
    'runzi/automation/arkiv',
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
